server/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.2
django-admin-interface==0.26.0
django-colorfield==0.8.0
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.0
frozenlist==1.3.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyproject_api==1.5.1
python-slugify==8.0.1
pytz==2023.3
redis==4.5.5
requests==2.31.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tox==4.6.0
twilio==8.3.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uuid==1.30
virtualenv==20.23.0
whitenoise==6.4.0
yarl==1.9.2
zope.interface==6.0
